PSGN10/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
django-import-export==2.4.0
django-rest-swagger==2.2.0
psycopg2-binary==2.8.6